pyproject.toml
[project]
name = "upstage-project-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "uvicorn[standard]>=0.38.0",
    "chromadb==1.3.7",
    "openai==1.52.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "httpx==0.27.1",
    "langgraph",
    "langchain-upstage",
    "langchain-chroma",
    "langchain-community",
    "langchain-core",
    "langsmith",
    "transformers",
    "streamlit",
    "pandas",
    "gdown",
]
